Update theme name and resources
The theme name is changed from "Anya-BD" to "Sailor-Moon-BD" and authorship has been updated. The CSS code has been revised to match the new theme, including changes to primary resources such as background and home button icons. This rebranding effort is part of the ongoing development to make the theme self-reliant.